LayoutTests/imported/w3c/web-platform-tests/css/cssom/resolved-border-width-expected.txt



FAIL checkbox assert_equals: border-width expected "5px" but got "0px"
FAIL radio assert_equals: border-width expected "5px" but got "0px"
